clean up app structure
