+++ .gitignore
@@ -1,3 +1,8 @@
-node_modules
-package-lock.json
+# exclude environment variables
 .env
+
+# exclude node libraries
+node_modules
+
+# exclude package-lock
+package-lock.json
