+++ app.js
@@ -1,0 +1,26 @@
+const express = require('express');
+const cors = require('cors');
+const cookieParser = require('cookie-parser');
+const logger = require('morgan');
+
+const indexRouter = require('./routes/index');
+const usersRouter = require('./routes/users');
+
+require('dotenv').config();
+
+const app = express();
+
+app.use(logger('dev'));
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
+app.use(cookieParser());
+app.use(cors());
+
+app.use('/', indexRouter);
+app.use('/users', usersRouter);
+
+const { PORT } = process.env;
+
+app.listen(PORT, () => console.log(`Server started on PORT ${PORT}`));
+
+module.exports = app;
