--- backend/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "data-collator-backend",
-  "version": "1.0.0",
-  "main": "src/index.js",
-  "dependencies": {
-    "bcrypt": "^5.0.0",
-    "bcryptjs": "^2.4.3",
-    "chai": "^4.3.0",
-    "chai-http": "4.3.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "mocha": "^8.3.2",
-    "mongoose": "^5.12.3",
-    "nodemon": "^2.0.6",
-    "uuid": "^8.3.2"
-  },
-  "devDependencies": {
-    "eslint": "^7.23.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx": "^0.1.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.23.1",
-    "eslint-plugin-react-hooks": "^4.2.0"
-  },
-  "scripts": {
-    "test": "mocha --require test/hooks.js --exit",
-    "serve": "nodemon src/index",
-    "lint": "eslint src --ext .js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": ""
-}
