--- backend/src/routes/.eslintrc.js
@@ -1,24 +1,0 @@
-module.exports = {
-  env: {
-    es6: true,
-    node: true,
-  },
-  extends: ['airbnb-base'],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly',
-  },
-  parserOptions: {
-    ecmaVersion: 2018,
-    sourceType: 'module',
-  },
-  rules: {
-    quotes: ['error', 'single', { allowTemplateLiterals: true }],
-    'consistent-return': 'off',
-    'class-methods-use-this': 'off',
-    'no-param-reassign': 'off',
-    'no-underscore-dangle': ['error', { allow: ['_id'] }],
-    'no-unused-vars': ['error', { argsIgnorePattern: 'next' }],
-    'no-console': 'off',
-  },
-};
