+++ package.json
@@ -1,0 +1,17 @@
+{
+  "name": "hackhate-backend",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "start": "node ./bin/www"
+  },
+  "dependencies": {
+    "cookie-parser": "~1.4.4",
+    "cors": "^2.8.5",
+    "debug": "~2.6.9",
+    "dotenv": "^9.0.0",
+    "express": "~4.16.1",
+    "mongoose": "^5.12.7",
+    "morgan": "~1.9.1"
+  }
+}
